SkyrimWeaponsAPI_Final/requirements.txt
Flask
Flask-SQLAlchemy
pandas
python-dotenv
